Extract some types exported from async-loaded that are referenced in sync-loaded modules into their own files.
